backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
pydantic-settings
email-validator
PyJWT
asyncpg
apscheduler
pytest
pytest-asyncio
httpx
